+++ controller/articles.controllers.js
@@ -1,4 +1,4 @@
-const { fetchArticleByID, updateArticleByID, fetchArticles, fetchArticleComments } = require("../model/articles.model")
+const { fetchArticleByID, updateArticleByID, fetchArticles } = require("../model/articles.model")
 
 
 exports.getArticleByID = (req, res, next) => {
@@ -29,12 +29,3 @@ exports.getArticles = (req, res, next) => {
     .catch(next)
 }
 
-exports.getArticleComments = (req, res, next) => {
-    const { article_id } = req.params;
-
-    const promises = [fetchArticleComments( article_id ), fetchArticleByID( article_id )]
-    Promise.all(promises).then(([comments]) => {
-        res.status(200).send({ comments })
-    })
-    .catch(next)
-}
